Save and Load system implemented, player movement mechanics balanced and ledge jumping bug fixes.
Bugs and Balancing:

Fixed bugs with jumping on the edge of ledges, now gives the player a grace period to keep jumping enabled after running a ground object.
Dash speed altered for mechanic balancing.
Devised a formula to increase player agility for a fixed player speed using absolute values and square roots of numbers between -1 and 1.
Adjusted player speed for game mechanic balancing.

New unintended mechanic discovered: Double Jump Wall Jump - If the feet transform object is placed to the left of the player, the circle detection clips with the wall, allowing the player to double jump from a wall jump. This was changed to make it only a single jump - feet transform at bottom middle of player collider.

Save and Load System:
Class created to store the state of game data.
Singleton class keeps track of the current state of the game. TODO: Enable new game saves. Currently saves and loads by default.
Interface that is implemented into other scripts to save and load data.
Serialization: File data handler class converts game data to a compressed format and writes it to a file. This format is currently JSON for ease of use, but could be updated to encrypted format to increase security in the future.
Deserialization: File data handler also reads the created file and converts the JSON data back to a unity C# object.

Tutorial followed by Trever Mock:
"How to make a Save & Load System in Unity | 2022"
https://github.com/trevermock

Objects:
Speed powerup was added as a test object, later to be removed.
Coins added to the game as testers for the dictionary storage. Each coin is given an ID so that it can be rendered or not, depending on whether the player has gathered the coin already or not. They will stay in the game. TODO: Add a second type of coin that is not recorded on a dictionary, but spawns at a random chance when an enemy is killed. This is because they are not "collectibles" but they are instead "currency" to be spent in-game at a later stage.
This coin collecting mechanic will be used in future for special collectibles and other common collectibles such as ammo, powerups and player lives.