Coins UI finished
Coins now increment in the player data script and save in the JSON file. The number of coins collected displays in the top corner.

TODO: either save the amount of coins spent by the player to remove when loading the game, or record the coins collected and the current amount of coins separately.